[IE TESTS][IE CONFORMANCE] Move Read_ir tests to Conformance
