+++ src/tests/functional/plugin/conformance/test_runner/api_conformance_runner/CMakeLists.txt
@@ -6,9 +6,9 @@ set(TARGET_NAME apiConformanceTests)
 
 addIeTargetTest(
         NAME ${TARGET_NAME}
-        ROOT "${CMAKE_CURRENT_SOURCE_DIR}"
+        ROOT "${CMAKE_CURRENT_SOURCE_DIR}/include"
         ADDITIONAL_SOURCE_DIRS
-            ${CMAKE_CURRENT_SOURCE_DIR}/src
+            "${CMAKE_CURRENT_SOURCE_DIR}/src"
         ADD_CPPLINT
         INCLUDES
             PRIVATE
