+++ src/tests/functional/plugin/conformance/test_runner/op_conformance_runner/CMakeLists.txt
@@ -6,9 +6,12 @@ set(TARGET_NAME conformanceTests)
 
 addIeTargetTest(
         NAME ${TARGET_NAME}
-        ROOT "${CMAKE_CURRENT_SOURCE_DIR}"
+        ROOT "${CMAKE_CURRENT_SOURCE_DIR}/include"
         ADDITIONAL_SOURCE_DIRS
-            ${CMAKE_CURRENT_SOURCE_DIR}/src
+            "${CMAKE_CURRENT_SOURCE_DIR}/src"
+        INCLUDES
+            PRIVATE
+                "${CMAKE_CURRENT_SOURCE_DIR}/include"
         ADD_CPPLINT
         LINK_LIBRARIES
             PUBLIC
