+++ src/tests/functional/plugin/conformance/test_runner/op_conformance_runner/include/read_ir_test/read_ir_compare_with_refs.hpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "shared_test_classes/read_ir/read_ir.hpp"
+#include "read_ir_test/read_ir.hpp"
 
 namespace ov {
 namespace test {
